+++ commons/src/test/java/com/karuslabs/commons/util/concurrent/locks/AutoReadWriteLockTest.java
@@ -113,4 +113,11 @@ class AutoReadWriteLockTest {
         verify(writer).isHeldByCurrentThread();
     }
     
+    @Test
+    void acquire() {
+        var lock = new AutoReadWriteLock();
+        assertNotNull(lock.readLock());
+        assertNotNull(lock.writeLock());
+    }
+    
 }